# cmd/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Bastio API
  description: AI Security Gateway API
  version: "0.1.0"
  license:
    name: FSL-1.1-ALv2
    url: https://fsl.software/FSL-1.1-ALv2.template.md
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /v1/traces:
    get:
      operationId: listTraces
      summary: List traces with filters
      tags: [Observability]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: end_user_id
          in: query
          schema:
            type: string
        - name: proxy_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Case-insensitive substring match on path.
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: environment
          in: query
          schema:
            type: string
        - name: release
          in: query
          schema:
            type: string
        - name: trace_name
          in: query
          schema:
            type: string
        - name: tag
          in: query
          description: key:value pair; repeat for multiple AND-ed tag filters.
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Paginated traces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trace"
  /v1/traces/{id}:
    get:
      operationId: getTrace
      summary: Get a trace with span hierarchy
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Trace detail with ordered spans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceDetail"
        "404":
          description: Trace not found
  /v1/traces/{id}/scores:
    get:
      operationId: listTraceScores
      summary: List scores attached to a trace
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scores for the trace (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TraceScore"
    post:
      operationId: createTraceScore
      summary: Attach a post-hoc score or label to a trace
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraceScoreRequest"
      responses:
        "201":
          description: Score recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceScore"
  /v1/traces/{id}/threats:
    get:
      operationId: listTraceThreats
      summary: Per-detector threat detections for a single trace
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Threat detections ordered by severity then score.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TraceThreatDetection"
  /v1/sessions:
    get:
      operationId: listSessions
      summary: List sessions (groups of traces sharing session_id)
      tags: [Observability]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: end_user_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: environment
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sessions ordered by most recent activity.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
  /v1/sessions/{id}:
    get:
      operationId: getSession
      summary: Session aggregates plus all traces in the session
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "404":
          description: Session not found
...
  /v1/users:
    get:
      operationId: listUsers
      summary: List end-users with activity rollups
      tags: [Observability]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: order_by
          in: query
          schema:
            type: string
            enum: [cost, requests, threats, latency]
            default: cost
      responses:
        "200":
          description: Aggregated per-end-user metrics.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAnalytics"
  /v1/users/{id}:
    get:
      operationId: getUser
      summary: End-user detail (sessions, recent traces, threat breakdown)
      tags: [Observability]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: User has no traces in this tenant.
  /v1/analytics/users:
    get:
      operationId: listUserAnalytics
      summary: Per-end-user cost / latency / threat rollups
      tags: [Analytics]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: order_by
          in: query
          schema:
            type: string
            enum: [cost, requests, threats, latency]
            default: cost
      responses:
        "200":
          description: Aggregated per-end-user metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAnalytics"
  /v1/threats:
    get:
      operationId: listThreats
      summary: List security threat events
      tags: [Security]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: severity
          in: query
          description: Exact-match severity (critical, high, medium, low).
          schema: { type: string }
        - name: threat_type
          in: query
          description: Exact-match threat category (jailbreak, pii, injection, bot, rate_anomaly, custom_pattern). Comma-separated for multi-value.
          schema: { type: string }
        - name: threat_subtype
          in: query
          description: Exact-match dotted sub-category (e.g. persona.dan, extraction.meta, structural.many_shot). Comma-separated for multi-value.
          schema: { type: string }
        - name: detector_name
          in: query
          description: Exact-match detector name that fired.
          schema: { type: string }
        - name: action_taken
          in: query
          description: Exact-match enforcement outcome (block, redact, warn, pass).
          schema: { type: string }
        - name: end_user_id
          in: query
          schema: { type: string }
        - name: ip_address
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: search
          in: query
          description: Case-insensitive substring match across matched_pattern and matched_content.
          schema: { type: string }
        - name: sort
          in: query
          description: Sort column. Unknown values fall back to detected_at.
          schema:
            type: string
            enum: [detected_at, severity, score, confidence]
            default: detected_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Threat events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatEvent"
  /v1/threats/{id}:
    get:
      operationId: getThreat
      summary: Get a single security threat event
      tags: [Security]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Threat event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatEvent"
        "404":
          description: Threat not found or belongs to another tenant
  /v1/analytics/overview:
    get:
      operationId: getAnalyticsOverview
      summary: Dashboard overview metrics
      tags: [Analytics]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Overview metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer sk-bastio-..."
  schemas:
...
    Trace:
      type: object
      required:
        - id
        - customer_id
        - proxy_id
        - method
        - path
        - provider
        - model
        - started_at
        - completed_at
        - duration_ms
        - input_tokens
        - output_tokens
        - total_tokens
        - cost_cents
        - status
        - threat_detected
        - threat_types
        - threat_score
        - security_action
        - end_user_id
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        proxy_id:
          type: string
          format: uuid
        method:
          type: string
        path:
          type: string
        provider:
          type: string
        model:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        total_tokens:
          type: integer
        cost_cents:
          type: number
        status:
          type: string
        threat_detected:
          type: boolean
        threat_types:
          type: array
          items:
            type: string
        threat_score:
          type: number
        security_action:
          type: string
        end_user_id:
          type: string
        session_id:
          type: string
        http_status:
          type: integer
        error_message:
          type: string
        request_body:
          type: string
          description: Populated by GET /v1/traces/{id}; empty in list responses.
        response_body:
          type: string
          description: Populated by GET /v1/traces/{id}; empty in list responses.
        environment:
          type: string
          description: Logical stage the request came from (production, staging, preview-pr-42, ...).
        release:
          type: string
          description: Application version or git SHA that produced the request.
        trace_name:
          type: string
          description: Semantic label for the request (e.g. guard_chat_completions).
        tags:
          type: object
          additionalProperties:
            type: string
          description: Free-form key/value tags attached to the trace via X-Bastio-Tag-* headers.
...
    TraceThreatDetection:
      type: object
      required: [id, trace_id, threat_type, severity, detector_name, detected_at]
      properties:
        id:
          type: string
          format: uuid
        trace_id:
          type: string
          format: uuid
        threat_type:
          type: string
          description: injection, pii, jailbreak, bot, rate_anomaly, custom_pattern
        severity:
          type: string
          enum: [low, medium, high, critical]
        score:
          type: number
        action_taken:
          type: string
        detector_name:
          type: string
        matched_pattern:
          type: string
        matched_content:
          type: string
        confidence:
          type: number
        end_user_id:
          type: string
        ip_address:
          type: string
        detected_at:
          type: string
          format: date-time
        threat_subtype:
          type: string
          description: Dotted sub-category emitted by the detector (e.g. persona.dan, extraction.meta).
        source:
          type: string
          description: Detection source label (e.g. pattern, structural, crescendo).
...
    SessionDetail:
      type: object
      required: [session, traces]
      properties:
        session:
          $ref: "#/components/schemas/Session"
        traces:
          type: array
          items:
            $ref: "#/components/schemas/Trace"
    TraceDetail:
      type: object
      required: [trace, spans]
      properties:
        trace:
          $ref: "#/components/schemas/Trace"
        spans:
          type: array
          items:
            $ref: "#/components/schemas/Observation"
    TraceScore:
      type: object
      required: [id, trace_id, name, value_type, created_at]
      properties:
        id:
          type: string
          format: uuid
        trace_id:
          type: string
          format: uuid
        name:
          type: string
        value_type:
          type: string
          enum: [numeric, categorical, boolean]
        numeric_value:
          type: number
          nullable: true
        string_value:
          type: string
        comment:
          type: string
        evaluator:
          type: string
        created_at:
          type: string
          format: date-time
    CreateTraceScoreRequest:
      type: object
      required: [name, value_type]
      properties:
        name:
          type: string
        value_type:
          type: string
          enum: [numeric, categorical, boolean]
        numeric_value:
          type: number
        string_value:
          type: string
        comment:
          type: string
        evaluator:
          type: string
...